Refactored test cases
